reorganizing post processing; break predict/evaluate out
