+++ Snakefile
@@ -3,7 +3,9 @@ import os
 # add scripts dir to path
 
 from river_dl.preproc_utils import prep_data
-from river_dl.postproc_utils import predict_from_weights, combined_metrics, plot_obs
+from river_dl.evaluate import combined_metrics
+from river_dl.postproc_utils import plot_obs
+from river_dl.predict import predict_from_weights
 from river_dl.train import train_model
 
 out_dir = config['out_dir']
